// src/app/models/ApplyTeachModel.ts
export interface ApplyTeachModel {
    
    DoB?: string;
    locations?: string;
    phone_num?: string;
    last_name?: string;
    phone?: string;
    location?: string;
    discipline?: string;  
    curriculum?:string;

    cv?: FormData;
    
    sp_ques_1?: string;    
    sp_ques_2?: string;     
    criminal_record?: string;  
    work_1?:string,
    work_1_detail?:string,
    condition?:string;

}